chore(core): restructure modules-related files
