+++ src/bp/bootstrap.ts
@@ -8,10 +8,10 @@ import sdk from 'botpress/sdk'
 import chalk from 'chalk'
 import cluster from 'cluster'
 import { Botpress, Config, Db, Ghost, LocalActionServer, Logger } from 'core/app'
-import { ModuleConfigEntry } from 'core/config/botpress.config'
+import { ModuleConfigEntry } from 'core/config'
 import { centerText } from 'core/logger'
-import { ModuleLoader } from 'core/module-loader'
-import ModuleResolver from 'core/modules/resolver'
+import { ModuleLoader, ModuleResolver } from 'core/modules'
+
 import fs from 'fs'
 import _ from 'lodash'
 import os from 'os'
