+++ src/bp/core/api.ts
@@ -7,6 +7,7 @@ import { SessionIdFactory } from 'core/dialog/sessions'
 import { EventEngine, EventRepository } from 'core/events'
 import { KeyValueStore } from 'core/kvs'
 import { LoggerProvider } from 'core/logger'
+import { ModuleLoader } from 'core/modules'
 import { WellKnownFlags } from 'core/sdk/enums'
 import { TYPES } from 'core/types'
 import { inject, injectable } from 'inversify'
@@ -21,7 +22,6 @@ import modelIdService from 'nlu-core/model-id-service'
 import { container } from './app.inversify'
 
 import { getMessageSignature } from './misc/security'
-import { ModuleLoader } from './module-loader'
 import { UserRepository } from './repositories'
 import { Event, RealTimePayload } from './sdk/impl'
 import HTTPServer from './server'
