+++ src/bp/core/app.inversify.ts
@@ -1,15 +1,16 @@
 import { Logger } from 'botpress/sdk'
+import { ConfigProvider } from 'core/config'
 import { EventCollector } from 'core/events'
+import { ModuleLoader } from 'core/modules'
 import { TelemetryContainerModules } from 'core/telemetry'
+import { TYPES } from 'core/types'
 import { Container } from 'inversify'
 
 import { BotpressAPIProvider } from './api'
 import { Botpress } from './botpress'
-import { ConfigProvider } from './config/config-loader'
 import { DatabaseContainerModules } from './database/database.inversify'
 import { LoggerDbPersister, LoggerFilePersister, LoggerProvider, PersistedConsoleLogger } from './logger'
 import { applyDisposeOnExit, applyInitializeFromConfig } from './misc/inversify'
-import { ModuleLoader } from './module-loader'
 import { RepositoriesContainerModules } from './repositories/repositories.inversify'
 import HTTPServer from './server'
 import { LocalActionServer } from './services/action/local-action-server'
@@ -19,7 +20,6 @@ import { DataRetentionService } from './services/retention/service'
 import { ServicesContainerModules } from './services/services.inversify'
 import { WorkspaceService } from './services/workspace-service'
 import { Statistics } from './stats'
-import { TYPES } from './types'
 
 const container = new Container({ autoBindInjectable: true })
