+++ src/bp/core/botpress.ts
@@ -3,10 +3,13 @@ import lang from 'common/lang'
 import { CMSService } from 'core/cms'
 import { BotpressConfig, ConfigProvider } from 'core/config'
 import { buildUserKey, converseApiEvents } from 'core/converse'
+import Database from 'core/database'
+import { StateManager } from 'core/dialog'
 import { SessionIdFactory } from 'core/dialog/sessions'
 import { addStepToEvent, EventCollector, StepScopes, StepStatus, EventEngine } from 'core/events'
 import { LoggerDbPersister, LoggerFilePersister, LoggerProvider, LogsJanitor } from 'core/logger'
 import { copyDir } from 'core/misc/pkg-fs'
+import { ModuleLoader } from 'core/modules'
 import { StatsService } from 'core/telemetry'
 import { WrapErrorsWith } from 'errors'
 import fse from 'fs-extra'
@@ -22,12 +25,7 @@ import plur from 'plur'
 
 import { startLocalActionServer } from '../cluster'
 import { setDebugScopes } from '../debug'
-
 import { createForGlobalHooks } from './api'
-
-import Database from './database'
-import { StateManager } from './dialog/state-manager'
-import { ModuleLoader } from './module-loader'
 import { WellKnownFlags } from './sdk/enums'
 import { Event } from './sdk/impl'
 import HTTPServer from './server'
