+++ src/bp/core/modules/index.ts
@@ -1,0 +1,6 @@
+export * from './module-loader'
+export * from './config-reader'
+export * from './resources-loader'
+export * from './utils/resolver'
+export * from './utils/require'
+export * from './modules-router'
