+++ src/bp/core/modules/utils/require.ts
