+++ src/bp/core/routers/admin/index.ts
@@ -2,9 +2,9 @@ import axios from 'axios'
 import { Logger } from 'botpress/sdk'
 import { checkRule } from 'common/auth'
 import LicensingService from 'common/licensing-service'
-import { ConfigProvider } from 'core/config/config-loader'
+import { ConfigProvider } from 'core/config'
 import { LogsRepository } from 'core/logger'
-import { ModuleLoader } from 'core/module-loader'
+import { ModuleLoader } from 'core/modules'
 import { GhostService } from 'core/services'
 import { AlertingService } from 'core/services/alerting-service'
 import AuthService, { TOKEN_AUDIENCE } from 'core/services/auth/auth-service'
