+++ src/bp/core/routers/admin/languages.ts
@@ -1,8 +1,8 @@
 import axios from 'axios'
 import { Logger } from 'botpress/sdk'
 import { StandardError, UnexpectedError } from 'common/http'
-import { ConfigProvider } from 'core/config/config-loader'
-import { ModuleLoader } from 'core/module-loader'
+import { ConfigProvider } from 'core/config'
+import { ModuleLoader } from 'core/modules'
 import { WorkspaceService } from 'core/services/workspace-service'
 import { RequestHandler, Router } from 'express'
 import Joi from 'joi'
