+++ src/bp/core/routers/admin/server.ts
@@ -1,11 +1,10 @@
 import { Logger } from 'botpress/sdk'
-import { ConfigProvider } from 'core/config/config-loader'
-import { ModuleLoader } from 'core/module-loader'
+import { ConfigProvider } from 'core/config'
+import { ModuleLoader } from 'core/modules'
 import { GhostService } from 'core/services'
 import { AlertingService } from 'core/services/alerting-service'
 import { JobService } from 'core/services/job-service'
 import { MonitoringService } from 'core/services/monitoring'
-import { WorkspaceService } from 'core/services/workspace-service'
 import diag from 'diag'
 import { Router } from 'express'
 import fse from 'fs-extra'
