+++ src/bp/core/routers/index.ts
@@ -1,4 +1,3 @@
-export * from './modules'
 export * from './bots'
 export * from './admin'
 export * from './auth'
