+++ src/bp/core/server.ts
@@ -10,6 +10,7 @@ import { CMSService, ContentRouter } from 'core/cms'
 import { ExternalAuthConfig, ConfigProvider } from 'core/config'
 import { ConverseRouter, ConverseService } from 'core/converse'
 import { LogsService, LogsRepository } from 'core/logger'
+import { ModuleLoader, ModulesRouter } from 'core/modules'
 import { TelemetryRouter, TelemetryRepository } from 'core/telemetry'
 import cors from 'cors'
 import errorHandler from 'errorhandler'
@@ -32,8 +33,7 @@ import portFinder from 'portfinder'
 import { URL } from 'url'
 import yn from 'yn'
 
-import { ModuleLoader } from './module-loader'
-import { AdminRouter, AuthRouter, BotsRouter, MediaRouter, ModulesRouter } from './routers'
+import { AdminRouter, AuthRouter, BotsRouter, MediaRouter } from './routers'
 import { HintsRouter } from './routers/bots/hints'
 import { NLURouter } from './routers/bots/nlu'
 import { isDisabled } from './routers/conditionalMiddleware'
