+++ src/bp/core/services/action/action-service.ts
@@ -6,9 +6,11 @@ import { ActionScope, ActionServer, LocalActionDefinition } from 'common/typings
 import { addErrorToEvent, addStepToEvent, StepScopes, StepStatus } from 'core/events'
 import { UntrustedSandbox } from 'core/misc/code-sandbox'
 import { printObject } from 'core/misc/print'
+import { clearRequireCache, requireFromString } from 'core/modules'
 import { TasksRepository } from 'core/repositories/tasks'
 import { NotFoundError } from 'core/routers/errors'
 import { ACTION_SERVER_AUDIENCE } from 'core/routers/sdk/utils'
+import { TYPES } from 'core/types'
 import { inject, injectable, tagged } from 'inversify'
 import joi from 'joi'
 import jsonwebtoken from 'jsonwebtoken'
@@ -20,12 +22,9 @@ import yn from 'yn'
 
 import { GhostService } from '..'
 import { createForAction } from '../../api'
-import { clearRequireCache, requireFromString } from '../../modules/require'
-import { TYPES } from '../../types'
 import { BotService } from '../bot-service'
 import { ActionExecutionError } from '../dialog/errors'
 import { WorkspaceService } from '../workspace-service'
-
 import { extractMetadata } from './metadata'
 import {
   enabled,
