+++ src/bp/core/services/action/utils.ts
@@ -1,6 +1,6 @@
 import { BUILTIN_MODULES } from 'common/defaults'
 import { ActionScope } from 'common/typings'
-import { requireAtPaths } from 'core/modules/require'
+import { requireAtPaths } from 'core/modules'
 import path from 'path'
 
 export const getBaseLookupPaths = (fullPath: string, lastPathPart: string) => {
