+++ src/bp/core/services/auth/auth-service.ts
@@ -8,7 +8,7 @@ import { SessionIdFactory } from 'core/dialog/sessions'
 import { EventEngine } from 'core/events'
 import { KeyValueStore } from 'core/kvs'
 import { getMessageSignature } from 'core/misc/security'
-import { ModuleLoader } from 'core/module-loader'
+import { ModuleLoader } from 'core/modules'
 import { BadRequestError } from 'core/routers/errors'
 import { Event } from 'core/sdk/impl'
 import { Response } from 'express'
