+++ src/bp/core/services/bot-service.ts
@@ -9,7 +9,7 @@ import { PersistedConsoleLogger } from 'core/logger'
 import { IDisposable } from 'core/misc/disposable'
 import { listDir } from 'core/misc/list-dir'
 import { stringify } from 'core/misc/utils'
-import { ModuleLoader } from 'core/module-loader'
+import { ModuleResourceLoader, ModuleLoader } from 'core/modules'
 import { RealTimePayload } from 'core/sdk/impl'
 import { Statistics } from 'core/stats'
 import { TYPES } from 'core/types'
@@ -35,7 +35,6 @@ import { FileContent, GhostService } from './ghost/service'
 import { Hooks, HookService } from './hook/hook-service'
 import { JobService } from './job-service'
 import { MigrationService } from './migration'
-import { ModuleResourceLoader } from './module/resources-loader'
 import RealtimeService from './realtime'
 import { WorkspaceService } from './workspace-service'
