+++ src/bp/core/services/dialog/flow/service.ts
@@ -3,7 +3,7 @@ import { ObjectCache } from 'common/object-cache'
 import { TreeSearch, PATH_SEPARATOR } from 'common/treeSearch'
 import { FlowMutex, FlowView, NodeView } from 'common/typings'
 import { KeyValueStore } from 'core/kvs'
-import { ModuleLoader } from 'core/module-loader'
+import { ModuleLoader } from 'core/modules'
 import { RealTimePayload } from 'core/sdk/impl'
 import { BotService } from 'core/services/bot-service'
 import RealtimeService from 'core/services/realtime'
