+++ src/bp/core/services/hook/hook-service.ts
@@ -4,6 +4,7 @@ import { ActionScope } from 'common/typings'
 import { addErrorToEvent, addStepToEvent, StepScopes, StepStatus } from 'core/events'
 import { UntrustedSandbox } from 'core/misc/code-sandbox'
 import { printObject } from 'core/misc/print'
+import { clearRequireCache, requireAtPaths } from 'core/modules'
 import { inject, injectable, tagged } from 'inversify'
 import _ from 'lodash'
 import ms from 'ms'
@@ -11,7 +12,6 @@ import path from 'path'
 import { NodeVM } from 'vm2'
 
 import { GhostService } from '..'
-import { clearRequireCache, requireAtPaths } from '../../modules/require'
 import { TYPES } from '../../types'
 import { filterDisabled, getBaseLookupPaths, runOutsideVm } from '../action/utils'
 import { VmRunner } from '../action/vm'
